fix admin panel

+++ database/migrations/2025_07_30_041626_add_slug_to_announcements_table.php
@@ -10,9 +10,11 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::table('announcements', function (Blueprint $table) {
-            $table->string('slug')->unique()->after('title');
-        });
+        if (!Schema::hasColumn('announcements', 'slug')) {
+            Schema::table('announcements', function (Blueprint $table) {
+                $table->string('slug')->unique()->after('title');
+            });
+        }
     }
 
     /**
@@ -20,8 +22,10 @@ return new class extends Migration {
      */
     public function down(): void
     {
-        Schema::table('announcements', function (Blueprint $table) {
-            $table->dropColumn('slug');
-        });
+        if (Schema::hasColumn('announcements', 'slug')) {
+            Schema::table('announcements', function (Blueprint $table) {
+                $table->dropColumn('slug');
+            });
+        }
     }
 };
