+++ database/migrations/2025_08_01_065915_add_slug_to_announcements_table.php
@@ -4,16 +4,17 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::table('announcements', function (Blueprint $table) {
-            $table->string('slug')->nullable()->after('title');
-        });
+        if (!Schema::hasColumn('announcements', 'slug')) {
+            Schema::table('announcements', function (Blueprint $table) {
+                $table->string('slug')->nullable()->after('title');
+            });
+        }
     }
 
     /**
@@ -21,8 +22,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('announcements', function (Blueprint $table) {
-            $table->dropColumn('slug');
-        });
+        if (Schema::hasColumn('announcements', 'slug')) {
+            Schema::table('announcements', function (Blueprint $table) {
+                $table->dropColumn('slug');
+            });
+        }
     }
 };
